Update README.md
One AI to rule them all.... This is an advanced AI powered coding system. When I give my AI a task, I want it to be able to execute shell scripts as needed. I want it to be able to use other AI language models as needed. Work in progress.